requirements.txt
pycrypto
beautifulsoup4
django
requests
PyJWT
django-cors-headers
mysqlclient